NFC: create GC interface for third-party GCs
